rename package to avoid varname shadowing
