+++ src/pkg/errs/err.go
@@ -1,4 +1,4 @@
-package err
+package errs
 
 import (
 	"errors"
