+++ src/pkg/errs/errs_test.go
@@ -1,4 +1,4 @@
-package err
+package errs
 
 import (
 	"testing"
